Return from _update if layer has been removed
When the layer is removed, this._map is set to null, which makes the
_update function fail if called. Since there is nothing to update when
the layer is removed, just return in this case.

The reason _update may be called after the layer has been removed is
that it uses a throttle so if the throttle is called e.g. two times
right after each other, the second call will be delayed and the layer
can be removed in that time. This can happen if you add the layer and
then remove it right after.